Implemented shield, missile base construction, and upgrading of bases.  Also added population bonus from soil and advanced soil enrichment.  Added setting output to zero when projects are completed.

+++ Beyond Beyaan/Data Modules/Galaxy.cs
@@ -704,7 +704,7 @@ namespace Beyond_Beyaan
 					{
 						owner = int.Parse(planetElement.Attribute("Owner").Value);
 					}
-					float populationMax = float.Parse(planetElement.Attribute("MaxPopulation").Value);
+					int populationMax = int.Parse(planetElement.Attribute("MaxPopulation").Value);
 					var newPlanet = new Planet(planetName, type, populationMax, null, newStar, gameMain.Random);
 					newPlanet.OwnerID = owner;
 					newPlanet.Factories = float.Parse(planetElement.Attribute("Buildings").Value);
